Rename detail/smart_holder_poc.h -> struct_smart_holder.h
